test: cover latest per series queries
