+++ src/query/src/dist_plan/planner.rs
@@ -14,7 +14,6 @@
 
 //! [ExtensionPlanner] implementation for distributed planner
 
-use std::any::Any;
 use std::sync::Arc;
 
 use ahash::HashMap;
@@ -25,18 +24,13 @@ use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME};
 use common_telemetry::debug;
 use datafusion::common::Result;
 use datafusion::datasource::DefaultTableSource;
-use datafusion::execution::context::{SessionState, TaskContext};
-use datafusion::physical_plan::metrics::MetricsSet;
-use datafusion::physical_plan::sorts::sort_preserving_merge::SortPreservingMergeExec;
-use datafusion::physical_plan::{
-    DisplayAs, DisplayFormatType, ExecutionPlan, PlanProperties, SendableRecordBatchStream,
-    Statistics,
-};
+use datafusion::execution::context::SessionState;
+use datafusion::physical_plan::ExecutionPlan;
 use datafusion::physical_planner::{ExtensionPlanner, PhysicalPlanner};
 use datafusion_common::tree_node::{TreeNode, TreeNodeRecursion, TreeNodeVisitor};
 use datafusion_common::{DataFusionError, TableReference};
 use datafusion_expr::{LogicalPlan, UserDefinedLogicalNode};
-use datafusion_physical_expr::{Distribution, LexOrdering, OrderingRequirements, PhysicalSortExpr};
+use datafusion_physical_expr::{LexOrdering, PhysicalSortExpr};
 use partition::manager::{PartitionRuleManagerRef, create_partitions_from_region_routes};
 use session::context::QueryContext;
 use snafu::{OptionExt, ResultExt};
@@ -47,7 +41,7 @@ use table::table_name::TableName;
 
 use crate::dist_plan::PredicateExtractor;
 use crate::dist_plan::merge_scan::{MergeScanExec, MergeScanLogicalPlan};
-use crate::dist_plan::merge_sort::MergeSortLogicalPlan;
+use crate::dist_plan::merge_sort::{MergeSortExec, MergeSortLogicalPlan};
 use crate::dist_plan::region_pruner::ConstraintPruner;
 use crate::error::{CatalogSnafu, PartitionRuleManagerSnafu, TableNotFoundSnafu};
 use crate::region_query::RegionQueryHandlerRef;
@@ -62,116 +56,6 @@ use crate::region_query::RegionQueryHandlerRef;
 ///
 /// We should ensure the number of partition is not smaller than the number of region at present. Otherwise this would result in incorrect output.
 pub struct MergeSortExtensionPlanner {}
-
-/// An opaque physical execution node for [`MergeSortLogicalPlan`].
-///
-/// It delegates execution and physical properties to DataFusion's
-/// [`SortPreservingMergeExec`], but intentionally does not expose itself as a
-/// `SortPreservingMergeExec`. `EnforceSorting` is allowed to replace a bare
-/// `SortPreservingMergeExec` with `CoalescePartitionsExec` when the parent does
-/// not require ordering. `MergeSortExec` represents the distributed TopK merge
-/// stage itself, so later physical optimizer rules must not rewrite it into an
-/// unordered fetch.
-#[derive(Debug, Clone)]
-struct MergeSortExec {
-    inner: SortPreservingMergeExec,
-}
-
-impl MergeSortExec {
-    fn new(ordering: LexOrdering, input: Arc<dyn ExecutionPlan>, fetch: Option<usize>) -> Self {
-        Self {
-            inner: SortPreservingMergeExec::new(ordering, input).with_fetch(fetch),
-        }
-    }
-}
-
-impl DisplayAs for MergeSortExec {
-    fn fmt_as(&self, t: DisplayFormatType, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match t {
-            DisplayFormatType::Default | DisplayFormatType::Verbose => {
-                write!(f, "MergeSortExec: [{}]", self.inner.expr())?;
-                if let Some(fetch) = self.inner.fetch() {
-                    write!(f, ", fetch={fetch}")?;
-                }
-                Ok(())
-            }
-            DisplayFormatType::TreeRender => self.inner.fmt_as(t, f),
-        }
-    }
-}
-
-impl ExecutionPlan for MergeSortExec {
-    fn name(&self) -> &str {
-        "MergeSortExec"
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn properties(&self) -> &Arc<PlanProperties> {
-        self.inner.properties()
-    }
-
-    fn required_input_distribution(&self) -> Vec<Distribution> {
-        self.inner.required_input_distribution()
-    }
-
-    fn benefits_from_input_partitioning(&self) -> Vec<bool> {
-        self.inner.benefits_from_input_partitioning()
-    }
-
-    fn required_input_ordering(&self) -> Vec<Option<OrderingRequirements>> {
-        self.inner.required_input_ordering()
-    }
-
-    fn maintains_input_order(&self) -> Vec<bool> {
-        self.inner.maintains_input_order()
-    }
-
-    fn children(&self) -> Vec<&Arc<dyn ExecutionPlan>> {
-        self.inner.children()
-    }
-
-    fn with_new_children(
-        self: Arc<Self>,
-        mut children: Vec<Arc<dyn ExecutionPlan>>,
-    ) -> Result<Arc<dyn ExecutionPlan>> {
-        Ok(Arc::new(Self::new(
-            self.inner.expr().clone(),
-            children.swap_remove(0),
-            self.inner.fetch(),
-        )))
-    }
-
-    fn execute(
-        &self,
-        partition: usize,
-        context: Arc<TaskContext>,
-    ) -> Result<SendableRecordBatchStream> {
-        self.inner.execute(partition, context)
-    }
-
-    fn metrics(&self) -> Option<MetricsSet> {
-        self.inner.metrics()
-    }
-
-    fn partition_statistics(&self, partition: Option<usize>) -> Result<Statistics> {
-        self.inner.partition_statistics(partition)
-    }
-
-    fn fetch(&self) -> Option<usize> {
-        self.inner.fetch()
-    }
-
-    fn with_fetch(&self, limit: Option<usize>) -> Option<Arc<dyn ExecutionPlan>> {
-        Some(Arc::new(Self::new(
-            self.inner.expr().clone(),
-            Arc::clone(self.inner.input()),
-            limit,
-        )))
-    }
-}
 
 impl MergeSortExtensionPlanner {
     fn ordering(
@@ -554,48 +438,5 @@ impl TreeNodeVisitor<'_> for TableNameExtractor {
             }
             _ => Ok(TreeNodeRecursion::Continue),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use datafusion::arrow::datatypes::{DataType, Field, Schema};
-    use datafusion::physical_expr::expressions::col as physical_col;
-    use datafusion::physical_plan::empty::EmptyExec;
-
-    use super::*;
-
-    #[test]
-    fn merge_sort_exec_is_opaque_and_preserves_topk_requirements() {
-        let schema = Arc::new(Schema::new(vec![Field::new("ts", DataType::Int64, false)]));
-        let input = Arc::new(EmptyExec::new(schema.clone()).with_partitions(2)) as _;
-        let ordering = LexOrdering::new([PhysicalSortExpr::new(
-            physical_col("ts", schema.as_ref()).unwrap(),
-            SortOptions {
-                descending: true,
-                nulls_first: false,
-            },
-        )])
-        .unwrap();
-
-        let merge_sort =
-            Arc::new(MergeSortExec::new(ordering, input, Some(1))) as Arc<dyn ExecutionPlan>;
-
-        assert_eq!(merge_sort.name(), "MergeSortExec");
-        assert!(
-            merge_sort
-                .as_any()
-                .downcast_ref::<SortPreservingMergeExec>()
-                .is_none(),
-            "MergeSortExec must stay opaque to EnforceSorting's bare SortPreservingMerge rewrite"
-        );
-        assert_eq!(merge_sort.fetch(), Some(1));
-        assert!(merge_sort.required_input_ordering()[0].is_some());
-
-        let fetched = merge_sort.with_fetch(Some(2)).unwrap();
-        assert!(fetched.as_any().downcast_ref::<MergeSortExec>().is_some());
-        assert_eq!(fetched.fetch(), Some(2));
     }
 }
